requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
qrcode==7.4.2
Pillow>=10.0.0
python-decouple==3.8
djangorestframework-simplejwt 